refactor tests for pystiche.loss (#369)
* refactor tests for pystiche.loss

* cleanup
